Add max_radius_mm and switch to polar coordinates for arc generation.

+++ GcodeProcessorLib/logger.cpp
@@ -24,7 +24,7 @@
 logger::logger(std::vector<std::string> names, std::vector<int> levels) {
 	// set to true by default, but can be changed by inheritance to support mandatory innitialization (for python or other integrations)
 	loggers_created_ = true;
-	num_loggers_ = names.size();
+	num_loggers_ = static_cast<int>(names.size());
 	logger_names_ = new std::string[static_cast<int>(num_loggers_)];
 	logger_levels_ = new int[static_cast<int>(num_loggers_)];
 	// this is slow due to the vectors, but it is trivial.  Could switch to an iterator
